Fix factories
Your attributes need to be lower case.

Your validation for postal_code length was failing because your column
was an integer instead of a string. This mean that the 0 in the
beginning was being stripped out. 02113 was turning into 2113, (wrong
length)

You can also use the `is` instead of min and max that are the same.