Collection select for choosing owner
You can use the collection select helper to create a select tag that
has an option for every Owner in your database. The additional
arguments being passed are to say to use the `id` of each Owner as the
value for each select and use the `full name` value for each Owner as
the label for each option. The prompt option lets us allow them to
optionally select a value.

More info at
http://apidock.com/rails/ActionView/Helpers/FormOptionsHelper/collection
_select